requirements.txt
asgiref==3.5.0
django-cors-headers==3.11.0
django-rest-framework==0.1.0
django-debugtools==2.0
Django==4.0.3
djangorestframework==3.13.1
pip==22.0.4
psycopg2==2.9.3
pytz==2022.1
setuptools==58.1.0
sqlparse==0.4.2
tzdata==2021.5
wheel==0.37.1
